Can now accelerate based on a vector.

+++ src/PhysicsBase/Vectors/VelocityVector.java
@@ -19,7 +19,7 @@ public class VelocityVector extends VectorBase
     }
 
     //SetMethods
-    public void SetSpeed(int speed)
+    public void SetSpeed(double speed)
     {
         _speed = speed;
     }
